requirements.txt
Bio
scikit-learn
cython
numpy
scipy
pysam
pyBigWig
pyVCF
fisher
HTSeq
hmmlearn
configparser
matplotlib
natsort
moods-python
pandas
logomaker
joblib